Create an OBJECTIVE-C-CLASS metaclass that can be used to
create CLOS classes that mirror Objective-C classes. Class
objects representing Objective-C classes contain a pointer
to the actual Objective-C class.

Creating an instance of one of these classes results in an
object with a FOREIGN-POINTER slot that points to an instance
of the corresponding Objective-C object. This object is created
by sending the "alloc" message to its Objective-C class.

The Objective-C object will have an Ivar that corresponds
to each slot specified in the DEFCLASS form. However,
syncing between the CLOS slots and the Objective-C Ivars
is not implemented. A substitute for SLOT-VALUE is needed,
since it cannot be overridden to do the needed syncing.

Also still TODO: Support Objective-C class-level Ivars.

Also still TODO: Go back over the code and raise errors
where the Objective-C runtime returns error codes.