Fix inheritance.
At this point, it's possible to:

- Create Objective-C classes and then subclass them.
- Access Objective-C IVARs using the Lisp SLOT-VALUE function.

There is some support for accessing integer IVARs (which requires
looking up the offset of the integer within the Objective-C object
and then dereferencing the object), but I have not exposed a reliable
way to specify the need to do this (or automatically detect it) in
the DEFCLASS form.

The OBJECTIVE-C-DIRECT-SLOT-DEFINITION class should be cleaned up.
It doesn't need many of the slots that it has. Removing the
unneeded slots would make room for the option to specify whether
an offset should be used for slot access.

Assigning an object using the OBJECTIVE-C-CLASS metaclass to
a slot expecting an Objective-C object does not work, because
the slot is expecting the raw pointer, not the CLOS object
that's wrapping it.